Improve temperature smoothing duration logic
Renamed TemperatureSmoothingDuration to MaximumTemperatureSmoothingDuration.
That value will define the maximum possible duration of the smoothing process, which will occur in those cases when the temperature delta is equal to the entire temperature range. Depending on how big or small the delta is, the smoothing process will take place slower or faster.